feat: 사용자 포인트 차감, 환불 기능 추가
fix: 타입 불일치 수정

+++ AuctionMarket/src/main/java/com/portfolio/auctionmarket/domain/auctions/entity/Auction.java
@@ -23,10 +23,10 @@ public class Auction {
     private Product product;
 
     @Column(name = "start_price")
-    private Integer startPrice;
+    private Long startPrice;
 
     @Column(name = "current_price")
-    private Integer currentPrice;
+    private Long currentPrice;
 
     @Column(name = "start_time")
     private LocalDateTime startTime;
@@ -38,7 +38,7 @@ public class Auction {
     @Enumerated(EnumType.STRING)
     private AuctionStatus status;
 
-    public void updateCurrentPrice(Integer currentPrice) {
+    public void updateCurrentPrice(Long currentPrice) {
         this.currentPrice = currentPrice;
     }
